requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-crontab==0.7.1
